Fix bug #64938: libxml_disable_entity_loader setting is shared between threads
Assigned CVE-IDs: CVE-2015-8866

The availability of entity loading is stored in a module global which
previously was only initialized in the GINIT constructor. This had the
effect that disabling the entity loader in one request caused
subsequent requests hitting the same Apache child process to  also have
the loader disabled.

With this change the loader is explicitely enabled in the request init
phase.

(cherry picked from commit de31324c221c1791b26350ba106cc26bad23ace9)
Signed-off-by: Lior Kaplan <kaplanlior@gmail.com>